samples: radio_test: Fix string formatting
These string formats were producing warnings.

config.mode is an enum and should therefore be formatted as %d.

channel_start and channel_end are unit8_t and should therefore not use
the l modifier.

It's not clear why lu was triggering a warning as the type is
uint32_t. Perhaps the correct modifier would have been llu, but u also
works for uint32_t.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>